refactor(@tinacms/tinacms): address review — require CollectionSchema.format (drop the ?? 'md' default), split markdown/json adapters into their own files, centralize the content slice's boundary cast in one documentsOf helper, and expand the provider boot-error ternary into if/else
