+++ packages/v4/@tinacms/tinacms/src/plugins/content/local/format-adapters.ts
@@ -1,8 +1,9 @@
-import matter from 'gray-matter';
 import type {
   CollectionSchema,
   TinaDocument,
 } from '../../../core/schema/types';
+import { jsonAdapter } from './json.adapter';
+import { markdownAdapter } from './markdown.adapter';
 
 export type CollectionFormat = NonNullable<CollectionSchema['format']>;
 
@@ -17,29 +18,6 @@ export interface FormatAdapter {
   parse(raw: string): TinaDocument;
   serialize(document: TinaDocument, previousRaw?: string): string;
 }
-
-// Frontmatter is the document; the body stays out of it until a rich-text `isBody`
-// field exists to own it — preserved via previousRaw meanwhile.
-const markdownAdapter = (extension: string): FormatAdapter => ({
-  extension,
-  parse: (raw) => matter(raw).data,
-  serialize: (document, previousRaw) => {
-    const previous = matter(previousRaw ?? '');
-    return matter.stringify(previous.content, {
-      ...previous.data,
-      ...document,
-    });
-  },
-});
-
-const jsonAdapter: FormatAdapter = {
-  extension: '.json',
-  parse: (raw) => JSON.parse(raw),
-  serialize: (document, previousRaw) => {
-    const previous = previousRaw ? JSON.parse(previousRaw) : {};
-    return `${JSON.stringify({ ...previous, ...document }, null, 2)}\n`;
-  },
-};
 
 const adapters: Partial<Record<CollectionFormat, FormatAdapter>> = {
   md: markdownAdapter('.md'),
