+++ packages/v4/@tinacms/tinacms/src/plugins/content/local/local-data-layer.ts
@@ -64,7 +64,7 @@ const resolveCollections = ({
     );
     resolved.set(schema.name, {
       schema,
-      adapter: formatAdapterFor(schema.format ?? 'md', formatAdapters),
+      adapter: formatAdapterFor(schema.format, formatAdapters),
       absoluteFolder: path.resolve(rootDir, schema.path),
     });
   }
